Move Redis connection options to a new file.
